# demoapi/api/openapi.yaml
openapi: 3.0.3
info:
  title: demoapi
  version: 1.0.0
  description: |
    A fake API to build TUIs against. Applications shaped like Argo CD's,
    jobs that run against them, AWX-shaped job events, and an
    Elasticsearch-shaped log index. See the demoapi package doc.

    Paging, filtering and sorting are plain parameters. Nothing here
    declares how a client should page: that belongs to the client (for a
    tuilib app, its design file), not to the spec.

    Every operation honours the failure knobs ?latency=, ?fail= and
    ?flaky=, applied before routing.

    It is a fixture. It has no authentication. Do not deploy it.
servers:
  - url: http://127.0.0.1:8099
paths:
  /apps:
    parameters:
      - $ref: "#/components/parameters/latency"
      - $ref: "#/components/parameters/fail"
      - $ref: "#/components/parameters/flaky"
    get:
      operationId: listApps
      summary: List applications, filtered, sorted and windowed.
      description: |
        Offset paging with the total in the body. `q` takes pkg/query's
        grammar whole; a parameter named after a column scopes to it. Both
        are AND-ed.
      parameters:
        - {name: q, in: query, schema: {type: string}, description: "pkg/query filter grammar: bare substrings, key:value column scopes, ~regex."}
        - {name: name, in: query, schema: {type: string}}
        - {name: region, in: query, schema: {type: string}}
        - {name: sync, in: query, schema: {type: string}}
        - {name: health, in: query, schema: {type: string}}
        - {name: phase, in: query, schema: {type: string}}
        - name: sort
          in: query
          description: Column to sort by (Name, Region, Sync, Health, Phase). Any prefix of a title, case-insensitive.
          schema: {type: string}
        - {name: desc, in: query, schema: {type: boolean}}
        - {name: offset, in: query, schema: {type: integer, minimum: 0, default: 0}}
        - {name: limit, in: query, schema: {type: integer, minimum: 0, default: 100}, description: "0 means every remaining row."}
        - $ref: "#/components/parameters/stale"
      responses:
        "200":
          description: One window of the filtered, sorted set.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AppPage"}
        default: {$ref: "#/components/responses/Error"}
  /apps/facets:
    parameters:
      - $ref: "#/components/parameters/latency"
      - $ref: "#/components/parameters/fail"
      - $ref: "#/components/parameters/flaky"
    get:
      operationId: listFacets
      summary: Every distinct value of a column, across the whole set.
      parameters:
        - name: field
          in: query
          required: true
          description: Column title or prefix.
          schema: {type: string}
      responses:
        "200":
          description: The distinct values, sorted.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Facets"}
        "400": {$ref: "#/components/responses/Error"}
        default: {$ref: "#/components/responses/Error"}
...
  /apps/{id}/{action}:
    parameters:
      - $ref: "#/components/parameters/appID"
      - name: action
        in: path
        required: true
        schema: {type: string, enum: [sync, fail]}
      - $ref: "#/components/parameters/latency"
      - $ref: "#/components/parameters/fail"
      - $ref: "#/components/parameters/flaky"
    post:
      operationId: launchJob
      summary: Start an operation against an application.
      description: Fire-and-observe. Answers 202 with a job id at once; the app's phase says Running.
      parameters:
        - {name: reconcile, in: query, schema: {type: string}, description: "Delay before the app reports the job. Duration or bare milliseconds."}
        - {name: blackhole, in: query, schema: {type: boolean}, description: "Accept, return an id, start nothing."}
        - {name: say, in: query, schema: {type: string}, description: "The phase the app reports while the job runs."}
        - name: phases
          in: query
          description: Phases walked in order across the job's duration.
          style: form
          explode: false
          schema: {type: array, items: {type: string}}
        - {name: takes, in: query, schema: {type: string}, description: "How long the work runs. Duration or bare milliseconds."}
      responses:
        "202":
          description: Accepted.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Launched"}
        "404": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
        default: {$ref: "#/components/responses/Error"}
  /jobs:
    parameters:
      - $ref: "#/components/parameters/latency"
      - $ref: "#/components/parameters/fail"
      - $ref: "#/components/parameters/flaky"
    get:
      operationId: listJobs
      summary: Jobs, newest first.
      parameters:
        - {name: app, in: query, schema: {type: string}, description: "Only this application's jobs."}
        - name: status
          in: query
          schema: {$ref: "#/components/schemas/JobStatus"}
      responses:
        "200":
          description: The jobs.
          content:
            application/json:
              schema:
                type: array
                items: {$ref: "#/components/schemas/Job"}
        default: {$ref: "#/components/responses/Error"}
...
  /chaos/down:
    post:
      operationId: chaosDown
      summary: Refuse every read for a window. Exempt from the outage itself.
      parameters:
        - name: for
          in: query
          description: How long. Duration or bare milliseconds; 0 ends an outage early.
          schema: {type: string}
      responses:
        "200":
          description: The outage is set.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Down"}
        default: {$ref: "#/components/responses/Error"}
  /chaos/delete/{id}:
    parameters:
      - $ref: "#/components/parameters/appID"
    post:
      operationId: chaosDelete
      summary: Remove an application from the set.
      responses:
        "200":
          description: Removed.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Deleted"}
        "404": {$ref: "#/components/responses/Error"}
        default: {$ref: "#/components/responses/Error"}
  /chaos/spawn:
    post:
      operationId: chaosSpawn
      summary: Add an application.
      parameters:
        - {name: id, in: query, required: true, schema: {type: string}, description: "May reuse an id something else had."}
        - {name: busy, in: query, schema: {type: boolean}, description: "Arrive already working."}
      responses:
        "201":
          description: Added.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/App"}
        "400": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
        default: {$ref: "#/components/responses/Error"}
components:
  parameters:
    latency:
      name: latency
      in: query
      description: Hold the request this long before answering. Duration or bare milliseconds.
      schema: {type: string}
    fail:
      name: fail
      in: query
      description: Answer with this status code (>= 400) instead.
      schema: {type: integer}
    flaky:
      name: flaky
      in: query
      description: Probability (0–1) of answering 502.
      schema: {type: number}
    stale:
      name: stale
      in: query
      description: Read as a replica lagging this far behind would. Duration or bare milliseconds.
      schema: {type: string}
    appID:
      name: id
      in: path
      required: true
      schema: {type: string}
...
  responses:
    Error:
      description: Any failure.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: {type: string}
    App:
      type: object
      description: One application, shaped the way Argo CD reports one.
      required: [id, name, region, sync, health, phase, reconciledAt, rev]
      properties:
        id: {type: string}
        name: {type: string}
        region: {type: string}
        sync:
          $ref: "#/components/schemas/SyncStatus"
        health: {type: string, description: "Healthy, Progressing or Degraded."}
        phase:
          type: string
          description: |
            The last operation's phase: empty before any, Running while one
            runs, then Succeeded or Failed. Not an enum: ?say= and Vocab put
            arbitrary values here on purpose.
        reconciledAt: {type: string, format: date-time}
        rev:
          type: integer
          format: int64
          description: Moves whenever a job against this app completes.
    SyncStatus:
      type: string
      description: A comparison, never an activity.
      enum: [Synced, OutOfSync]
    AppPage:
      type: object
      required: [total, offset, rows]
      properties:
        total: {type: integer, description: "Rows matching the filter, across every page."}
        offset: {type: integer}
        rows:
          type: array
          items: {$ref: "#/components/schemas/App"}
    Facets:
      type: object
      required: [values]
      properties:
        values:
          type: array
          items: {type: string}
...
    JobKind:
      type: string
      enum: [sync, fail]
    JobStatus:
      type: string
      enum: [running, succeeded, failed]
    Job:
      type: object
      required: [id, app, kind, status, started]
      properties:
        id: {type: string}
        app: {type: string}
        kind: {$ref: "#/components/schemas/JobKind"}
        status: {$ref: "#/components/schemas/JobStatus"}
        started: {type: string, format: date-time}
        finished: {type: string, format: date-time, description: "Absent while running."}
    Launched:
      type: object
      required: [job]
      properties:
        job: {type: string, description: "The job's id, for GET /jobs/{id}."}
...
    Down:
      type: object
      required: [down_for]
      properties:
        down_for: {type: string}
    Deleted:
      type: object
      required: [deleted]
      properties:
        deleted: {type: string}